Move common functions to a utilities file for easier reuse
